refactor: rename api-v2 to api, use Flex for sidebar
- Merge method-badge and json-editor into api folder
- Remove api-v2 directory, update all imports
- Use Flex components for API sidebar nodes
