--- packages/chronicle/src/components/api-v2/index.ts
@@ -1,5 +1,0 @@
-export { ApiOverview } from './api-overview'
-export { ApiFieldSection } from './api-field-list'
-export { ApiCodeSnippet } from './api-code-snippet'
-export { ApiResponsePanel } from './api-response-panel'
-export { PlaygroundDialog } from './playground-dialog'
